organization: split expression template code into chunks, began adding documentation
